Some refactoring done, improve button now has changing text
