+++ lib/customWidgets/BreathingImage.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
-import 'package:useful_app/blocs/HomeScreen/HomeScreenBloc.dart';
+import 'package:useful_app/util/WidgetValues.dart';
 
 class BreathingImage extends StatefulWidget {
